requirements.txt
pandas
pydantic
pyyaml